+++ backend/src/main/java/linklibrary/repository/UserRepository.java
@@ -6,4 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
+
+    public User findByLoginId(String loginId);
 }
