+++ backend/src/main/java/linklibrary/security/auth/PrincipalDetailService.java
@@ -1,4 +1,25 @@
 package linklibrary.security.auth;
 
-public class PrincipalDetailService {
+import linklibrary.entity.User;
+import linklibrary.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class PrincipalDetailService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        log.info("loadUserByUsername 실행");
+        User userEntity = userRepository.findByLoginId(username);
+        return new PrincipalDetails(userEntity);
+    }
 }
