+++ backend/src/main/java/linklibrary/security/auth/PrincipalDetails.java
@@ -18,6 +18,7 @@ public class PrincipalDetails implements UserDetails {
     }
 
     //사용자가 가지고 있는 권한 정보를 반환
+    //권한을 하나만 갖고 있기 때문에 singleton 객체 반환
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return Collections.singleton(new SimpleGrantedAuthority(user.getRole().toString())); //역할을 문자열로 바꾸어 return
@@ -25,31 +26,31 @@ public class PrincipalDetails implements UserDetails {
 
     @Override
     public String getPassword() {
-        return null;
+        return user.getPassword();
     }
 
     @Override
     public String getUsername() {
-        return null;
+        return user.getLoginId();
     }
 
     @Override
     public boolean isAccountNonExpired() {
-        return false;
+        return true;
     }
 
     @Override
     public boolean isAccountNonLocked() {
-        return false;
+        return true;
     }
 
     @Override
     public boolean isCredentialsNonExpired() {
-        return false;
+        return true;
     }
 
     @Override
     public boolean isEnabled() {
-        return false;
+        return true;
     }
 }
